requirements.txt
streamlit==1.31.1
requests==2.31.0
numpy==1.24.3
pandas==2.1.4
arxiv==2.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0 